Bring aws-c-compression into the 20th century

+++ source/huffman_generator/generator.c
@@ -200,7 +200,7 @@ void huffman_node_write_decode(struct huffman_node *node, FILE *file, uint8_t cu
     write_label = 1;
 
     /* Check 1 bit pattern */
-    uint32_t single_bit_mask = 1ull << (31 - current_bit);
+    uint32_t single_bit_mask = (uint32_t)(1ull << (31 - current_bit));
     uint32_t left_aligned_pattern = ((node->code.bits << 1) + 1) << (31 - node->code.num_bits);
     uint32_t check_pattern = left_aligned_pattern & single_bit_mask;
     fprintf(file, "    if (bits & 0x%x) {\n", check_pattern);
